Updated all automated date fields
`date_created` should not be nullable
`date_updated` should not be nullable. The initial insert is just a
special "update" so it can be filled into this field as well.

Changed absolute `pg_dump` path to be relative.
Added a `manipulate-schema.ts` that should act as a starting point if
you want to do programatic manipulation of the yaml schema for mass
changes.