chore: change update script to be aggressive (see comment)
I believe this is a more sane default for directus because we very rarely run this script. Theres a very real risk of quickly falling behind in the js ecosystem. And so that rare time we do update directus, we should invest by being a bit aggressive and willing to take the time to fix whatever breaking changes arise.